added more ray tracing stuff

+++ math/plane.c
@@ -1,7 +1,7 @@
 #include "./plane.h"
 #include "vector3d.h"
 
-int plane_intersects(Ray *ray, void *data)
+float plane_intersects(Ray *ray, void *data)
 {
     Plane* p = (Plane*)data;
     float fv = dot(&p->normal, &ray->direction);
